新增预约接口

+++ app/Activity/Modules/Repository/ActivityGoodsAppointmentRepository.php
@@ -16,6 +16,11 @@ class ActivityGoodsAppointmentRepository
     /**
      * 创建活动与商品的关联关系
      * @param $data
+     * [
+     *    'appointment_id'  =>'',  //活动id   int  【必传】
+     *    'spu_id'          =>'',  //商品id   int  【必传】
+     *    'create_time'    =>'',  //创建时间  int  【必传】
+     * ]
      * @return bool
      */
     public static function add(array $data){
@@ -31,6 +36,9 @@ class ActivityGoodsAppointmentRepository
      * @return array
      */
     public static function getByIdInfo(int $id){
+        if(empty($id)){
+            return false;
+        }
         $where[]=['appointment_id','=',$id];
         $activityInfo=ActivityGoodsAppointment::where($where)->get()->toArray();
         if( !$activityInfo ){
@@ -40,7 +48,17 @@ class ActivityGoodsAppointmentRepository
 
     }
 
+    /***
+     * 删除活动和商品的关系数据
+     * @param int $id  活动id
+     * @return bool|null
+     * @throws \Exception
+     */
+
     public static function delActivityGoods(int $id){
+        if(empty($id)){
+            return false;
+        }
         $where[]=['appointment_id','=',$id];
         $res=ActivityGoodsAppointment::where($where)->delete();
         if( !$res ){
@@ -48,24 +66,7 @@ class ActivityGoodsAppointmentRepository
         }
         return $res;
     }
-    /***
-     * 获取活动信息
-     * @return array
-     */
-    public static  function getActivityInfo(){
 
-    }
-    /***
-     * 执行编辑活动
-     * @param $data
-     * [
-     * 'spu_id'    =>['',''] 商品id     int      【必传】
-     * ]
-     * @return bool
-     */
-    public static function activityUpdate(array $data){
-
-    }
 
 
 }
